requirements.txt
streamlit==1.30.0 
tensorflow==2.15.0
Pillow==10.0.1
numpy==1.24.3
keras==2.15.0